requirements.txt
--find-links wheelhouse
django >=1.9, <2
django-tables2 >=1.2
pyodbc==3.0.10
django-pyodbc-azure >=1.9